% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llreg.R
\name{llreg}
\alias{llreg}
\title{log-log regression}
\usage{
llreg(df, x, y)
}
\arguments{
\item{df}{data frame containing the variables of interest}

\item{x}{predictor / explanatory variable, passed in quotes}

\item{y}{response variable, passed in quotes}
}
\value{
elasticity, intercept, standard error, and p-value estimates from the regression output
}
\description{
A quick way to estimate elasticity between 2 variables.
}
\examples{
llreg(mtcars, 'mpg', 'cyl')
}
