% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctchg.R
\name{pctchg}
\alias{pctchg}
\title{Calculating Percentage Changes}
\usage{
pctchg(df, col)
}
\arguments{
\item{df}{dataframe containing column of interest}

\item{col}{name of the column, passed in quotes}
}
\value{
a new dataframe with an additional column containing "pctchg_"
}
\description{
\code{pctchg} takes a dataframe and a numeric column in it to calculate percentage changes from one period to the next. If values are missing/0, the function will return 0 for that observation in the "pctchg_" column. Time series dataframes recommended.
}
\examples{
pctchg(treedat, 'x')
}
