% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_3pac.R
\name{theme_3pac}
\alias{theme_3pac}
\title{A Clean ggplot2 Theme}
\usage{
theme_3pac(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{numerical value, the size of the plot's fonts, default value is 12}

\item{base_family}{the family of fonts the graph is based on, default value is 'sans'}
}
\value{
an additional function within a ggplot() call
}
\description{
\code{theme_3pac} is a function based on \code{theme_bw}, meant to improve the visual clarity of \code{ggplot} objects.
}
\examples{
library(ggplot2)
ggplot(data = treedat, aes(x = x, y = y, col = z)) +
  geom_point() +
  geom_smooth(se = FALSE) +
  labs(title = 'Example scatter with best fit',
       subtitle = 'theme_3pac() example',
       caption = 'Data were randomly generated and are not reflective of any real-life trend') +
  theme_3pac()
}
