% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedat.R
\docType{data}
\name{treedat}
\alias{treedat}
\title{A toy dataset to demonstrate the functions of the treepac library}
\format{
\subsection{\code{treedat}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{time}{An integer denoting the time period from 1 to 100}
\item{x}{Dummy explanatory variable / predictor}
\item{y}{Dummy response variable whose natural log is a function of the natural log of x}
}
}
}
\usage{
treedat
}
\description{
A toy dataset to demonstrate the functions of the treepac library
}
\keyword{datasets}
