% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedat.R
\docType{data}
\name{treedat}
\alias{treedat}
\title{Treedat}
\format{
\subsection{\code{treedat}}{

A data frame with 100 rows and 4 columns:
\describe{
\item{period}{An integer denoting the time period from 1 to 100}
\item{x}{Dummy explanatory variable / predictor}
\item{y}{Dummy response variable whose natural log is a function of the natural log of x}
\item{z}{Dummy categorical variable}
}
}
}
\usage{
treedat
}
\description{
A dataframe to motivate the treepac's available functions.
}
\keyword{datasets}
